+++ vscode/src/core/parsers/PackageJsonParser.ts
@@ -1,16 +1,14 @@
 import { Settings } from "../../config";
-import * as vscode from "vscode";
 import Item from "../Item";
 import { JsonParser } from "./JsonParser";
 import { PackageLockJsonParser } from "./PackageLockJsonParser";
-import path, { dirname } from "path";
-import * as fs from "fs";
-import { clearText } from "./TomlParser";
 
 export class NpmParser extends JsonParser {
   constructor() {
     super(
       [
+        "catalog",
+        "catalogs",
         "dependencies",
         "devDependencies",
         "peerDependencies",
@@ -22,8 +20,11 @@ export class NpmParser extends JsonParser {
     );
   }
 
+  
+
   addDependency(item: Item) {
-    if (item.value?.startsWith("link:")) {
+    const skipPrefixes = ["link:", "catalog:", "workspace:"];
+    if (skipPrefixes.some(prefix => item.value?.startsWith(prefix))) {
       return;
     }
     if (item.value?.startsWith("jsr:") || item.value?.startsWith("npm:@jsr/")) {
@@ -32,55 +33,7 @@ export class NpmParser extends JsonParser {
     item = convertAliasToPackageName(item);
     item.createRange();
     item.createDecoRange();
-    if (item.value?.startsWith("catalog:")) {
-      this.loadYamlLinesIfNeeded();
-
-      if (this.state.yamlLines.length > 0) {
-        this.updateItemValueFromCatalog(item);
-      }
-    }
     this.state.items.push(item);
-  }
-  private loadYamlLinesIfNeeded() {
-    if (this.state.yamlLines.length === 0) {
-      const filePath = vscode.window.activeTextEditor?.document.uri.fsPath;
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(
-        vscode.Uri.file(filePath || "")
-      );
-      const rootPath = workspaceFolder?.uri.fsPath as string;
-
-      try {
-        const files = fs.readdirSync(rootPath);
-        const yamlFile = files.find((file) => file === "pnpm-workspace.yaml");
-        if (yamlFile) {
-          const yamlFilePath = path.join(rootPath, yamlFile);
-          const yamlContent = fs.readFileSync(yamlFilePath, "utf8");
-          this.state.yamlLines = yamlContent.split("\n");
-        }
-      } catch (err) {
-        console.error(err);
-      }
-    }
-  }
-
-  private updateItemValueFromCatalog(item: Item) {
-    let catalogName = item.value?.replace("catalog:", "");
-    if (catalogName?.trim() === "") {
-      catalogName = item.key;
-    }
-
-    const catalogLine = this.state.yamlLines.find((line) => {
-      const catalogKey = line
-        .split(":")[0]
-        .trim()
-        .replace(/^['"]|['"]$/g, "");
-      return catalogKey === item.key || catalogKey === catalogName;
-    });
-
-    if (catalogLine) {
-      const catalogValue = clearText(catalogLine.split(":")[1].trim());
-      item.value = catalogValue;
-    }
   }
 }